bot/requirements.txt
dotenv
aiohttp
discord.py